Rename settings subfolder
